# backend/salonbw-backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/stocktaking": {
      "get": {
        "operationId": "StocktakingController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "draft",
                "in_progress",
                "completed",
                "cancelled"
              ],
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lista wszystkich inwentaryzacji",
        "tags": [
          "stocktaking"
        ]
      },
      "post": {
        "operationId": "StocktakingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStocktakingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Utwórz nową inwentaryzację",
        "tags": [
          "stocktaking"
        ]
      }
    },
    "/stocktaking/history": {
      "get": {
        "operationId": "StocktakingController_findHistorySummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Historia inwentaryzacji (agregaty)",
        "tags": [
          "stocktaking"
        ]
      }
    },
    "/stocktaking/{id}": {
      "get": {
        "operationId": "StocktakingController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pobierz inwentaryzację po ID",
        "tags": [
          "stocktaking"
        ]
      },
      "patch": {
        "operationId": "StocktakingController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStocktakingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aktualizuj inwentaryzację",
        "tags": [
          "stocktaking"
        ]
      }
    },
    "/stocktaking/{id}/start": {
      "post": {
        "operationId": "StocktakingController_start",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rozpocznij inwentaryzację (załaduj produkty)",
        "tags": [
          "stocktaking"
        ]
      }
    },
    "/stocktaking/{id}/items": {
      "post": {
        "operationId": "StocktakingController_addItems",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStocktakingItemsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dodaj/aktualizuj pozycje inwentaryzacji",
        "tags": [
          "stocktaking"
        ]
      }
    },
    "/stocktaking/{id}/items/{itemId}": {
      "patch": {
        "operationId": "StocktakingController_updateItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStocktakingItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aktualizuj pozycję inwentaryzacji",
        "tags": [
          "stocktaking"
        ]
      }
    },
    "/stocktaking/{id}/complete": {
      "post": {
        "operationId": "StocktakingController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteStocktakingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Zakończ inwentaryzację (opcjonalnie zastosuj różnice)",
        "tags": [
          "stocktaking"
        ]
      }
    },
...
  },
  "info": {
    "title": "SalonBW API",
    "description": "The SalonBW API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateStocktakingDto": {
        "type": "object",
        "properties": {
          "stocktakingDate": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "UpdateStocktakingDto": {
        "type": "object",
        "properties": {
          "stocktakingDate": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "StocktakingItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number"
          },
          "countedQuantity": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "productId",
          "countedQuantity"
        ]
      },
      "AddStocktakingItemsDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StocktakingItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "UpdateStocktakingItemDto": {
        "type": "object",
        "properties": {
          "countedQuantity": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "CompleteStocktakingDto": {
        "type": "object",
        "properties": {
          "applyDifferences": {
            "type": "boolean",
            "default": true
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
    }
  }
}